requirements.txt

pillow
rembg
gunicorn==20.1.0
onnxruntime
